Исправил ошибки

+++ back/metadata-service/Models/Transcript.cs
@@ -4,9 +4,9 @@ public sealed class Transcript
 {
     public long   Id           { get; init; }
     public Guid   SongId       { get; init; }
-    public int    SegmentIndex { get; init; }
-    public int    StartMs      { get; init; }
-    public int    EndMs        { get; init; }
+    public long    SegmentIndex { get; init; }
+    public long    StartMs      { get; init; }
+    public long    EndMs        { get; init; }
     public string Text         { get; set; } = null!;
     public decimal? Confidence { get; set; }
     public Guid?  CheckedById  { get; set; }
